Fix max power bug
There was an error in the flap direction where it would always have a large Z component. This meant that every flap was at maximum power, and was just partially into the screen. The result was that every flap cost the maximum amount, and the max power radius was not effective.

This revision fixes the bug and tweaks the max power settings appropriately.